requirements.txt
djangorestframework==3.13.1
django==3.2.0

